LINQ Literal comparison issue
Corrected an issue when LINQ processor compared two literals of type
System.Uri. Due to RFC 986 spec, System.Uri.Equals method treats Uris
with the hash (#) fragment as equal. In case of semantic web
technologies this behavior is not welcome as each URI different in a
single char should be considered as separate. This is somehow why we
introduced EntityId type in favour of the System.Uri