Dropped ITypedEntity casting for internal usage.
Library was using ITypedEntity for internal type checks/assertions. This
was correct from the architecture point of view, but created a lot of
additional calls which hit the performance. This is dropped in favour of
in-memory store calls outside entity proxies.

This also corrects some minor issues and adds support for file based
in-memory stores.